Add error messages

+++ api/request_response/trigger_chatbot.go
@@ -1,8 +1,8 @@
 package requestresponse
 
 type TriggerChatbot struct {
-	ChatroomID                  int  `json:"chatroom_id,omitempty"`
-	MessageID                   int  `json:"message_id,omitempty"`
-	ShouldStreamChatbotResponse bool `json:"should_stream_chatbot_response,omitempty"`
-	ApiVersion                  int  `json:"api_version,omitempty"`
+	ChatroomID                  int   `json:"chatroom_id,omitempty"`
+	MessageID                   int64 `json:"message_id,omitempty"`
+	ShouldStreamChatbotResponse bool  `json:"should_stream_chatbot_response,omitempty"`
+	ApiVersion                  int   `json:"api_version,omitempty"`
 }
